requirements.txt
boto3==1.34.4
datasets==3.6.0
gitpython==3.1.43
javalang==0.13.0
numpy==1.26.4
parameterized==0.8.1
protobuf==3.20.3
pydantic==2.6.4
pylint==2.14.5
pyspark==3.5.1
pytz==2024.1
packaging
requests==2.32.3
nltk
rank_bm25
